Move all indexers into same package
