+++ indexers/contacts.go
@@ -1,4 +1,4 @@
-package contacts
+package indexers
 
 import (
 	"bytes"
@@ -8,37 +8,30 @@ import (
 	"fmt"
 	"time"
 
-	indexer "github.com/nyaruka/rp-indexer"
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
 )
 
-//go:embed index_settings.json
-var indexSettings json.RawMessage
-
-// indexes a contact
-const indexCommand = `{ "index": { "_id": %d, "_type": "_doc", "version": %d, "version_type": "external", "routing": %d} }`
-
-// deletes a contact
-const deleteCommand = `{ "delete" : { "_id": %d, "_type": "_doc", "version": %d, "version_type": "external", "routing": %d} }`
+//go:embed contacts.settings.json
+var contactsSettings json.RawMessage
 
 // ContactIndexer is an indexer for contacts
-type Indexer struct {
-	indexer.BaseIndexer
+type ContactIndexer struct {
+	baseIndexer
 
 	batchSize int
 }
 
-// NewIndexer creates a new contact indexer
-func NewIndexer(elasticURL, name string, batchSize int) *Indexer {
-	return &Indexer{
-		BaseIndexer: indexer.NewBaseIndexer(elasticURL, name),
+// NewContactIndexer creates a new contact indexer
+func NewContactIndexer(elasticURL, name string, batchSize int) *ContactIndexer {
+	return &ContactIndexer{
+		baseIndexer: newBaseIndexer(elasticURL, name),
 		batchSize:   batchSize,
 	}
 }
 
 // Index indexes modified contacts and returns the name of the concrete index
-func (i *Indexer) Index(db *sql.DB, rebuild, cleanup bool) (string, error) {
+func (i *ContactIndexer) Index(db *sql.DB, rebuild, cleanup bool) (string, error) {
 	var err error
 
 	// find our physical index
@@ -54,11 +47,11 @@ func (i *Indexer) Index(db *sql.DB, rebuild, cleanup bool) (string, error) {
 
 	// doesn't exist or we are rebuilding, create it
 	if physicalIndex == "" || rebuild {
-		physicalIndex, err = i.CreateNewIndex(indexSettings)
+		physicalIndex, err = i.createNewIndex(contactsSettings)
 		if err != nil {
 			return "", errors.Wrap(err, "error creating new index")
 		}
-		i.Log().WithField("index", physicalIndex).Info("created new physical index")
+		i.log().WithField("index", physicalIndex).Info("created new physical index")
 		remapAlias = true
 	}
 
@@ -67,7 +60,7 @@ func (i *Indexer) Index(db *sql.DB, rebuild, cleanup bool) (string, error) {
 		return "", errors.Wrap(err, "error finding last modified")
 	}
 
-	i.Log().WithField("index", physicalIndex).WithField("last_modified", lastModified).Info("indexing newer than last modified")
+	i.log().WithField("index", physicalIndex).WithField("last_modified", lastModified).Info("indexing newer than last modified")
 
 	// now index our docs
 	start := time.Now()
@@ -76,11 +69,11 @@ func (i *Indexer) Index(db *sql.DB, rebuild, cleanup bool) (string, error) {
 		return "", errors.Wrap(err, "error indexing documents")
 	}
 
-	i.RecordComplete(indexed, deleted, time.Since(start))
+	i.recordComplete(indexed, deleted, time.Since(start))
 
 	// if the index didn't previously exist or we are rebuilding, remap to our alias
 	if remapAlias {
-		err := i.UpdateAlias(physicalIndex)
+		err := i.updateAlias(physicalIndex)
 		if err != nil {
 			return "", errors.Wrap(err, "error updating alias")
 		}
@@ -89,7 +82,7 @@ func (i *Indexer) Index(db *sql.DB, rebuild, cleanup bool) (string, error) {
 
 	// cleanup our aliases if appropriate
 	if cleanup {
-		err := i.CleanupIndexes()
+		err := i.cleanupIndexes()
 		if err != nil {
 			return "", errors.Wrap(err, "error cleaning up old indexes")
 		}
@@ -98,8 +91,72 @@ func (i *Indexer) Index(db *sql.DB, rebuild, cleanup bool) (string, error) {
 	return physicalIndex, nil
 }
 
+const sqlSelectModifiedContacts = `
+SELECT org_id, id, modified_on, is_active, row_to_json(t) FROM (
+	SELECT
+		id,
+		org_id,
+		uuid,
+		name,
+		language,
+		status,
+		ticket_count AS tickets,
+		is_active,
+		created_on,
+		modified_on,
+		last_seen_on,
+		EXTRACT(EPOCH FROM modified_on) * 1000000 AS modified_on_mu,
+		(
+			SELECT array_to_json(array_agg(row_to_json(u)))
+			FROM (SELECT scheme, path FROM contacts_contacturn WHERE contact_id = contacts_contact.id) u
+		) AS urns,
+		(
+			SELECT jsonb_agg(f.value)
+			FROM (
+				SELECT 
+					CASE
+					WHEN value ? 'ward'
+					THEN jsonb_build_object('ward_keyword', trim(substring(value ->> 'ward' from  '(?!.* > )([^>]+)')))
+					ELSE '{}'::jsonb
+					END || district_value.value AS value
+				FROM (
+					SELECT 
+						CASE
+						WHEN value ? 'district'
+						THEN jsonb_build_object('district_keyword', trim(substring(value ->> 'district' from  '(?!.* > )([^>]+)')))
+						ELSE '{}'::jsonb
+						END || state_value.value as value
+					FROM (
+						SELECT 
+							CASE
+							WHEN value ? 'state'
+							THEN jsonb_build_object('state_keyword', trim(substring(value ->> 'state' from  '(?!.* > )([^>]+)')))
+							ELSE '{}' :: jsonb
+							END || jsonb_build_object('field', key) || value as value
+						FROM jsonb_each(contacts_contact.fields)
+					) state_value
+				) AS district_value
+			) AS f
+		) AS fields,
+		(
+			SELECT array_to_json(array_agg(g.uuid)) FROM (
+				SELECT contacts_contactgroup.uuid
+				FROM contacts_contactgroup_contacts, contacts_contactgroup
+				WHERE contact_id = contacts_contact.id AND contacts_contactgroup_contacts.contactgroup_id = contacts_contactgroup.id
+			) g
+		) AS groups,
+		(
+			SELECT f.uuid FROM flows_flow f WHERE f.id = contacts_contact.current_flow_id
+		) AS flow
+	FROM contacts_contact
+	WHERE modified_on >= $1
+	ORDER BY modified_on ASC
+	LIMIT 500000
+) t;
+`
+
 // IndexModified queries and indexes all contacts with a lastModified greater than or equal to the passed in time
-func (i *Indexer) indexModified(db *sql.DB, index string, lastModified time.Time) (int, int, error) {
+func (i *ContactIndexer) indexModified(db *sql.DB, index string, lastModified time.Time) (int, int, error) {
 	batch := &bytes.Buffer{}
 	createdCount, deletedCount, processedCount := 0, 0, 0
 
@@ -111,7 +168,7 @@ func (i *Indexer) indexModified(db *sql.DB, index string, lastModified time.Time
 	start := time.Now()
 
 	for {
-		rows, err := FetchModified(db, lastModified)
+		rows, err := db.Query(sqlSelectModifiedContacts, lastModified)
 
 		queryCreated := 0
 		queryCount := 0
@@ -152,7 +209,7 @@ func (i *Indexer) indexModified(db *sql.DB, index string, lastModified time.Time
 
 			// write to elastic search in batches
 			if queryCount%i.batchSize == 0 {
-				created, deleted, err := i.IndexBatch(index, batch.Bytes())
+				created, deleted, err := i.indexBatch(index, batch.Bytes())
 				if err != nil {
 					return 0, 0, err
 				}
@@ -165,7 +222,7 @@ func (i *Indexer) indexModified(db *sql.DB, index string, lastModified time.Time
 		}
 
 		if batch.Len() > 0 {
-			created, deleted, err := i.IndexBatch(index, batch.Bytes())
+			created, deleted, err := i.indexBatch(index, batch.Bytes())
 			if err != nil {
 				return 0, 0, err
 			}
@@ -186,7 +243,7 @@ func (i *Indexer) indexModified(db *sql.DB, index string, lastModified time.Time
 		elapsed := time.Since(start)
 		rate := float32(processedCount) / (float32(elapsed) / float32(time.Second))
 
-		i.Log().WithField("index", index).WithFields(logrus.Fields{"rate": int(rate), "added": createdCount, "deleted": deletedCount, "elapsed": elapsed}).Info("indexed contact batch")
+		i.log().WithField("index", index).WithFields(logrus.Fields{"rate": int(rate), "added": createdCount, "deleted": deletedCount, "elapsed": elapsed}).Info("indexed contact batch")
 	}
 
 	return createdCount, deletedCount, nil
